Adds in-memory card data storage
At the moment the card data, which is used to determine how to render
each participant's ID on various pages, is loaded from disk and parsed
each time it's used.

This change adds an in-memory version of this data that is stored in an
efficient manner relating to how that data is used by the UI

In terms of timings, the complete refresh of the ids before the
check_legal lru cache is populated, takes 2min 49sec.
Repopulating via this new method takes roughly 300ms.

In terms of how this change will help general performance, one cause of
latency that was identified is that a portion of the load when
rendering the tournaments page appears to be coming from repeatedly
reading and parsing of the ids.json file from disk.

Every time `get_faction` is called, which is twice per player on the
tournament details page, the entire set of ids is loaded from disk,
parsed as json, and then searched through linearly looking for the
faction of the user's ids.

Switching this from a disk-read, parse, etc to a straight memory lookup
in the in-memory Cards class appears to have improved performance.

Some cursory load testing shows that performance degrades quite a bit
under concurrent reads, given each request will attempt to load the ids
file from disk. The change to in-memory, at least on the machine used to
develop the change, appears to cut the latency by about a third.

Concurrent Readers | Disk-Read Latency | In-Memory Latency |
1                  | 239ms             | 154ms             |
2                  | 631ms             | 323ms             |
5                  | 1655ms            | 1047ms            |
10                 | 3023ms            | 1991ms            |

This test was by no means exhaustive but gives an indication that
there's an improvement here.

An additional change here also pre-calculates the list of identities to
render in the UI drop downs during registration.

Currently when we render the list of IDs, both Runner and Corp, we
re-generate the lists every time that the render the form for a user to
register. This involves a lot of work to split the ids into legal and
non-legal lists, sort them, and then recombine them into the final list.
In addition to this, the original get_ids function is being called
multiple times, each of which will load the list of ids from disk and
parse them.

This change includes the generation of these Corp and Runner ID lists
into the reprocessing of card data from NetrunnerDB. This occurs only
once, and the result of that work can then be referenced each time we
need to render the form, rather than regenerating it every time.

In addition to this, a background task was added to periodically
refresh card data. When the cards module is loaded, we start a thread
to run in the background and update the card data every hour in a way
that doesn't add any extra load to the user who's request is running
when the card data expires.