Improve performance of the tournament page
This change includes additional performance improvements for the
tournament page, mostly related to efficient access to the database.

In the previous version, the players were initially ranked by 3 factors:
esos, sos, and score. The esos and sos scores are calculated as results
are recorded and fairly efficient. However one main driver of latency
here is including the score in the calculation by calling the player
business logic's 'get_record' function. This function iterates all of
the players results, both corp and runner, meaning that every time one
player was compared to another in the sorting algorithm, each player's
matches were iterated and evaluated again.

In addition to this, 'get_record' is also called again when rendering
the results page, 6 times each per player, which is another 6 iterations
of each players matches.

This change improves this by ensuring we only look at each match once,
and incrementally update the record for each player involved in the
match accordingly, before sorting them once at the end, using their
score value that we calculated during the single iteration of matches.

In addition to this, the calculation of the top cut rankings was called
3 times when rendering the page, so we make the optimization of only
calling it once, and reusing the results multiple times when rendering.

On the machine used to develop these changes, the 99%ile latency has
improved quite a bit on a tournament with 250 participants.

| Concurrent Reads | Prior Latency | New Latency |
| 1                | 158ms         | 39ms        |
| 2                | 343ms         | 56ms        |
| 5                | 1111ms        | 207ms       |
| 10               | 2175ms        | 479ms       |