Fixes an issue parsing decklists with right single quotation marks
There's an issue parsing decklists that include the right single
quotation mark character, such as "Knickknack" O’Brian.

In the format that we receive data from NetrunnerDB, the stripped title
has the right single quotation mark (’) normalized to a regular quotation
mark ('). The Jinteki.net format however includes the original right single
quotation mark, causing us to not be able to find the card correctly.

This change is a simple one to normalize the quotation mark on our end,
as well as adding an additional test to ensure this doesn't break in the
future.

Also cleans up some extraneous 'print' statements.

Fixes #46